Final project. API is complete , but needs to be fleshed out to handle edge cases and custom exceptions.

+++ Booking.WebApi/ViewModels/ClientViewModel.cs
@@ -8,7 +8,7 @@ namespace Booking.WebApi.ViewModels
 {
     public class ClientViewModel
     {
-        public uint Id { get; set; }
+        public byte Id { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public byte Age { get; set; }
